+++ components/pages/Dashboard/EditingLayer.tsx
@@ -4,13 +4,13 @@ import PencilIcon from '@/components/icons/pencil.svg';
 import TrashIcon from '@/components/icons/trash.svg';
 import { DashboardContext } from '@/components/pages/Dashboard/context';
 import { startAppStateLoadingTransition } from '@/core/bind-client';
-import { duplicateItem } from '@/core/helpers/items';
 import { MenuItem } from '@/packages/ui/components/menu';
 import { ToolbarMenu } from '@/packages/ui/components/toolbar-menu';
 import useRefCallback from '@/packages/ui/hooks/ref-callback';
-import { useDeleteDashboardItem } from '@/store/features/dashboards';
-import { useLibraryItemField, useUpdateLibraryItem } from '@/store/features/library';
+import dashboardsFeature, { useDeleteDashboardItem } from '@/store/features/dashboards';
+import libraryFeature, { useLibraryItemField, useUpdateLibraryItem } from '@/store/features/library';
 import { useWidget } from '@/store/features/widgets';
+import store from '@/store/store';
 import { getConfigurable, getDuplicable } from '@/utils/widgets';
 import EyeSlashIcon from 'bootstrap-icons/icons/eye-slash.svg';
 import EyeIcon from 'bootstrap-icons/icons/eye.svg';
@@ -57,7 +57,7 @@ export function EditingLayer ({ id }: EditLayerProps) {
   });
 
   const handleDuplicate = useRefCallback(() => {
-  
+
     duplicateItem(id);
   });
 
@@ -127,3 +127,37 @@ export function EditingLayer ({ id }: EditLayerProps) {
     </div>
   );
 }
+
+function cloneJson<T> (val: T): T {
+  if (val && typeof val === 'object') {
+    return JSON.parse(JSON.stringify(val));
+  }
+  return val;
+}
+
+function duplicateItem (id: string, props?: (props: any) => any) {
+  const { library, dashboards } = store.getState();
+
+  const item = library.items[id];
+  if (!dashboards.current) {
+    return;
+  }
+  const dashboard = dashboards.dashboards[dashboards.current];
+  const oldItem = dashboard.items[id];
+  if (dashboard && item && oldItem) {
+    const prev = item;
+    const prevProps = cloneJson(prev.props);
+    const newItem = {
+      id: `${prev.name}-${Math.round(Date.now() / 1000)}`,
+      name: prev.name,
+      props: cloneJson(props?.(prevProps) ?? prevProps),
+    };
+    const newReference = {
+      id: newItem.id,
+      layout: cloneJson(oldItem.layout),
+    };
+    store.dispatch(libraryFeature.actions.add({ item: newItem }));
+    store.dispatch(dashboardsFeature.actions.add({ item: newReference }));
+  }
+}
+
