--- core/layout/defaults.ts
@@ -1,6 +1,0 @@
-import { Dashboard } from '@/utils/types/config';
-
-export const defaultLayoutConfig: Dashboard['layout'] = {
-  size: [40, 16],
-  gap: 8,
-};
